+++ generators/model/Generator.php
@@ -181,6 +181,20 @@ class Generator extends \yii\gii\Generator
     }
 
     /**
+     * @return string db tables prefix
+     *
+     */
+    public function getTablePrefix()
+    {
+      $db = $this->getDbConnection();
+      if ($db !== null) {
+            return $db->tablePrefix;
+      } else {
+            return "";
+      }
+    }
+
+    /**
      * @inheritdoc
      */
     public function generate()